#52: Extract TextIndex
