+++ logic/src/main/java/org/itsallcode/whiterabbit/logic/autocomplete/AutocompleteService.java
@@ -1,6 +1,5 @@
 package org.itsallcode.whiterabbit.logic.autocomplete;
 
-import static java.util.Collections.emptyList;
 import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.counting;
 import static java.util.stream.Collectors.groupingBy;
@@ -8,13 +7,10 @@ import static java.util.stream.Collectors.toList;
 
 import java.time.LocalDate;
 import java.time.Period;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.SortedSet;
-import java.util.TreeSet;
 import java.util.stream.Stream;
 
 import org.apache.logging.log4j.LogManager;
@@ -42,12 +38,12 @@ public class AutocompleteService
 
     public AutocompleteEntrySupplier dayCommentAutocompleter()
     {
-        return autocompleter(getDayComments());
+        return TextIndex.build(getDayComments());
     }
 
     public AutocompleteEntrySupplier activityCommentAutocompleter()
     {
-        return autocompleter(getActivityComments());
+        return TextIndex.build(getActivityComments());
     }
 
     private List<String> getDayComments()
@@ -82,32 +78,6 @@ public class AutocompleteService
     {
         final LocalDate maxAge = clockService.getCurrentDate().minus(MAX_AGE);
         return storage.getLatestDays(maxAge);
-    }
-
-    AutocompleteEntrySupplier autocompleter(Collection<String> allEntries)
-    {
-        LOG.debug("Creating autocompleter for {} entries: {}", allEntries.size(), allEntries);
-        final Map<String, List<String>> lowerCaseIndex = allEntries.stream().collect(groupingBy(String::toLowerCase));
-        final SortedSet<String> lowerCaseValues = new TreeSet<>(lowerCaseIndex.keySet());
-        return searchText -> {
-            if (searchText == null || searchText.isBlank())
-            {
-                return emptyList();
-            }
-            final SortedSet<String> lowerCaseMatches = lowerCaseValues.subSet(searchText.toLowerCase(),
-                    searchText.toLowerCase() + Character.MAX_VALUE);
-            return lowerCaseMatches.stream()
-                    .map(lowerCaseIndex::get)
-                    .flatMap(List::stream)
-                    .map(proposedText -> createProposal(searchText, proposedText))
-                    .collect(toList());
-        };
-    }
-
-    private AutocompleteProposal createProposal(String searchText, String proposedText)
-    {
-        final int matchPositionStart = proposedText.toLowerCase().indexOf(searchText.toLowerCase());
-        return new AutocompleteProposal(proposedText, matchPositionStart, searchText.length());
     }
 
     public Optional<Project> getSuggestedProject()
